Renamed crate and library directories, updated <librender_cdk> folder structure.
