+++ librender_cdk/librender_cdk/lib.cpp
@@ -1,0 +1,6 @@
+#include "environment_manager.h"
+#include <iostream>
+
+int main() {
+  std::cout << "[API_KEY] -> " << EnvironmentManager::getApiKey() << std::endl;
+}
