+++ render_cdk/src/iaas/blueprints/paths.rs
