+++ render_cdk/src/resource_management/models/template.rs
